Refactor file paths and update dependencies
