+++ prisma/schema.prisma
@@ -54,8 +54,8 @@ model User {
   user_id     Int        @id @default(autoincrement())
   email       String     @unique
   password    String
-  role        Role
-  profile_id  Int
+  role        Role?
+  profile_id  Int?
   student     Student?
   teacher     Teacher?
   hod         Hod?
