+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import LandingPage from "@/components/landing/LandingPage";
+import LandingPage from "@/src/components/landing/LandingPage";
 
 export default function Home() {
   return (
