+++ src/components/magicui/marquee.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/lib/utils";
+import { cn } from "@/src/lib/utils";
 
 interface MarqueeProps {
   className?: string;
